+++ src/main/java/ch/opentrainingcenter/model/Weather.java
@@ -6,46 +6,46 @@ import javax.persistence.Id;
 @Entity
 public class Weather {
 
-	@Id
-	private int id;
-	private String wetter;
-	private String imageicon;
+    @Id
+    private long id;
+    private String wetter;
+    private String imageicon;
 
-	public Weather() {
-	}
+    public Weather() {
+    }
 
-	public Weather(final int id) {
-		this.id = id;
-	}
+    public Weather(final int id) {
+        this.id = id;
+    }
 
-	public int getId() {
-		return this.id;
-	}
+    public long getId() {
+        return this.id;
+    }
 
-	public void setId(final int id) {
-		this.id = id;
-	}
+    public void setId(final long id) {
+        this.id = id;
+    }
 
-	public String getWetter() {
-		return this.wetter;
-	}
+    public String getWetter() {
+        return this.wetter;
+    }
 
-	public void setWetter(final String wetter) {
-		this.wetter = wetter;
-	}
+    public void setWetter(final String wetter) {
+        this.wetter = wetter;
+    }
 
-	public String getImageicon() {
-		return this.imageicon;
-	}
+    public String getImageicon() {
+        return this.imageicon;
+    }
 
-	public void setImageicon(final String imageicon) {
-		this.imageicon = imageicon;
-	}
+    public void setImageicon(final String imageicon) {
+        this.imageicon = imageicon;
+    }
 
-	@Override
-	@SuppressWarnings("nls")
+    @Override
+    @SuppressWarnings("nls")
 
-	public String toString() {
-		return "Weather [wetter=" + wetter + ", imageicon=" + imageicon + "]";
-	}
+    public String toString() {
+        return "Weather [wetter=" + wetter + ", imageicon=" + imageicon + "]";
+    }
 }
