+++ src/main/java/ch/opentrainingcenter/service/RepositoryService.java
@@ -2,12 +2,12 @@ package ch.opentrainingcenter.service;
 
 public interface RepositoryService<T> {
 
-    T find(Class<T> type, int id);
+    T find(Class<T> type, long id);
 
     T doSave(T t);
 
     T update(T t);
 
-    void remove(Class<T> type, int id);
+    void remove(Class<T> type, long id);
 
 }
