+++ src/main/java/ch/opentrainingcenter/service/persistence/RepositoryServiceBean.java
@@ -10,29 +10,29 @@ import ch.opentrainingcenter.service.RepositoryService;
 @Stateless
 public class RepositoryServiceBean<T> implements RepositoryService<T> {
 
-	@PersistenceContext(unitName = "otc", type = PersistenceContextType.TRANSACTION)
-	protected EntityManager em;
+    @PersistenceContext(unitName = "otc", type = PersistenceContextType.TRANSACTION)
+    protected EntityManager em;
 
-	@Override
-	public T doSave(final T t) {
-		em.persist(t);
-		return t;
-	}
+    @Override
+    public T doSave(final T t) {
+        em.persist(t);
+        return t;
+    }
 
-	@Override
-	public T update(final T t) {
-		return em.merge(t);
-	}
+    @Override
+    public T update(final T t) {
+        return em.merge(t);
+    }
 
-	@Override
-	public void remove(final Class<T> type, final int id) {
-		final T managed = em.find(type, id);
-		em.remove(managed);
-	}
+    @Override
+    public void remove(final Class<T> type, final long id) {
+        final T managed = em.find(type, id);
+        em.remove(managed);
+    }
 
-	@Override
-	public T find(final Class<T> type, final int id) {
-		return em.find(type, id);
-	}
+    @Override
+    public T find(final Class<T> type, final long id) {
+        return em.find(type, id);
+    }
 
 }
